Set horizontal margins of wg_layer to 0. Renamed wg... to wg_...
